+++ src/UpForm/UpFormGroup.tsx
@@ -46,6 +46,7 @@ const UpFormGroup = (props: UpFormGroupProps) => {
       </>
     );
   }
+  
   return (
     <>
       {title && (